Bugfix positioning on left
When the element that opens the colorpicker is located at the right edge of
the screen a calculation is done and the colorpicker is shown on the left.
However it is not shown underneath and on the left but diagonal to the left

This bugfix fixes this behavior by subtracting the width of the element
that opened the color picker off the left positioning.